requirements.txt
openstacksdk
paramiko
pyyaml

